requirements.txt
Flask==3.0.0
python-telegram-bot==13.12
requests